+++ client/src/scripts/input/input.ts
@@ -5,9 +5,11 @@ import { createLogger } from "../logger";
 const log = createLogger("input");
 
 export function initInput(scene: GameScene, room: Room<GameState>) {
-  const { up, down, left, right } = scene.input.keyboard.createCursorKeys();
+  const { up, down, left, right, space } =
+    scene.input.keyboard.createCursorKeys();
 
   let lastDirection = { up: false, down: false, left: false, right: false };
+  let lastSprint = false;
   const updateListener = scene.events.on("update", () => {
     const curDirection = {
       up: up.isDown,
@@ -25,6 +27,16 @@ export function initInput(scene: GameScene, room: Room<GameState>) {
       // TODO further optimize by only sending changed directions
       room.send("direction", curDirection);
       log("sent new direction %j", curDirection);
+    }
+    if (lastSprint !== space.isDown) {
+      lastSprint = space.isDown;
+      if (space.isDown) {
+        log("sprinting");
+        room.send("sprint", true);
+      } else {
+        log("stop sprinting");
+        room.send("sprint", false);
+      }
     }
   });
 }