+++ client/src/scripts/objects/fpsText.ts
@@ -1,11 +1,14 @@
-export default class FpsText extends Phaser.GameObjects.Text {
+import { DEPTH } from "../globals";
+
+export default class GameText extends Phaser.GameObjects.Text {
   constructor(scene) {
-    super(scene, 10, 10, '', { color: 'black', fontSize: '28px' })
-    scene.add.existing(this)
-    this.setOrigin(0)
+    super(scene, 10, 10, "", { color: "black", fontSize: "28px" });
+    scene.add.existing(this);
+    this.setOrigin(0);
+    this.setDepth(DEPTH.text);
   }
 
   public update() {
-    this.setText(`fps: ${Math.floor(this.scene.game.loop.actualFps)}`)
+    this.setText(`Energy: ${Math.floor(this.scene.game.loop.actualFps)}`);
   }
 }